Encryption of email password added.
The key and the initialization vector are stored at the same machine which is potentially dangerous. Therefore, the password should not be set on the client, but rather on the server at a place where only the administrator has access to.

This is a premise for building a client-server architecture.